requirements/common.txt
requests==2.28.2
pandas==1.5.3
python-dotenv==1.0.0
azure-functions==1.13.3
azure-functions-durable==1.2.2
azure-cosmos == 4.3.1
colorhash == 1.2.1
jsonschema == 4.17.3
python-dateutil == 2.8.2
